+++ core/XH.js
@@ -21,6 +21,7 @@ import {
     GridExportService,
     IdentityService,
     IdleService,
+    InspectorService,
     JsonBlobService,
     LocalStorageService,
     PrefService,
@@ -60,6 +61,7 @@ class XHClass {
     #initCalled = false;
     #lastActivityMs = Date.now();
     #uaParser = null;
+    #services = [];
 
     constructor() {
         makeObservable(this);
@@ -114,6 +116,8 @@ class XHClass {
     identityService;
     /** @member {IdleService} */
     idleService;
+    /** @member {InspectorService} */
+    inspectorService;
     /** @member {JsonBlobService} */
     jsonBlobService;
     /** @member {LocalStorageService} */
@@ -264,6 +268,7 @@ class XHClass {
                 install the same service twice.`
             ));
             this[name] = svc;
+            this.#services.push(svc);
         });
     }
 
@@ -644,8 +649,8 @@ class XHClass {
     /**
      * Return a collection of models currently 'active' in this application.
      *
-     * This will include all models that have not had the destroy() method
-     * called on them.  Models will be returned in creation order.
+     * This will include all models that have not had their destroy() method called.
+     * Models will be returned in creation order.
      *
      * @param {ModelSelector} [selector] - optional selector for filtering models.
      * @returns {HoistModel[]}
@@ -656,6 +661,11 @@ class XHClass {
             if (m.matchesSelector(selector, true)) ret.push(m);
         });
         return ret;
+    }
+
+    /** @return {HoistService[]} - all services registered with this application. */
+    getServices() {
+        return [...this.#services];
     }
 
     /**
@@ -816,7 +826,7 @@ class XHClass {
 
             await this.installServicesAsync(
                 AlertBannerService, AutoRefreshService, ChangelogService, IdleService,
-                GridAutosizeService, GridExportService, WebSocketService
+                InspectorService, GridAutosizeService, GridExportService, WebSocketService
             );
             this.acm.init();